+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to the `dom_query` crate will be documented in this file.
 
+## [0.23.1] - 2025-10-12
+
+### Added
+- **Markdown**: `MDSerializer::find_code_language` now uses `MDSerializer::find_code_language_css_class` as a fallback when it fails to detect the `code` language from the ancestors (by @justahero).
+
+### Fixed
+- **Markdown**: Escape double quotes in `title` attributes of `<a>` elements (by @justahero).
+
+## [0.23.0] - 2025-09-28
+
+### Added
+- Implemented `Element::attr_ref` method which returns an `&str` reference to the attribute value by `html5ever::LocalName`.
+- Re-exported `html5ever::LocalName` and `html5ever::local_name`.
+- **Markdown**: Enhanced `<pre>`-block parsing by checking its associated `data-lang` and `data-language` attributes (by @justahero)
+- **Markdown**: Multiline `<code>` blocks are now parsed as `<pre>` blocks (by @justahero).
+### Changed
+- Revised `NodeRef::find_descendants` (requires `mini_selector` feature). This method now supports `Adjacent (+)` and `Sibling (~)` combinators.
+
+### Fixed
+- **Markdown**: fixed block element linebreak handling.
+
 ## [0.22.0] - 2025-09-03
 
 ### Added